Convert bias analyzer development script to component
